completed api for now

+++ src/main/java/com/wayfare/backend/controller/UserController.java
@@ -2,29 +2,27 @@ package com.wayfare.backend.controller;
 
 
 import java.time.Instant;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.wayfare.backend.model.BadgeEnum;
 import com.wayfare.backend.model.Review;
 import com.wayfare.backend.model.TourListing;
-import com.wayfare.backend.model.dto.UserDTO;
 import com.wayfare.backend.repository.ReviewRepository;
 import com.wayfare.backend.repository.TourRepository;
 import com.wayfare.backend.request.EditAccountDetailsRequest;
 import com.wayfare.backend.request.EditProfileRequest;
+import com.wayfare.backend.request.PasswordRequest;
+import com.wayfare.backend.response.AccountSettingsResponse;
 import com.wayfare.backend.response.ProfileResponse;
 import com.wayfare.backend.response.ResponseObject;
 import com.wayfare.backend.model.User;
 import com.wayfare.backend.repository.UserRepository;
 
 import com.wayfare.backend.security.WayfareUserDetails;
-import com.wayfare.backend.validator.AlphabeticalValidator;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.parameters.P;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 
 import static com.wayfare.backend.helper.helper.getCurrentUserDetails;
@@ -32,13 +30,15 @@ import static com.wayfare.backend.helper.helper.getCurrentUserDetails;
 @RestController
 public class UserController {
 
+    private final AuthenticationManager authenticationManager;
     private final UserRepository userRepo;
     private final ReviewRepository reviewRepo;
     private final TourRepository tourRepo;
 
     ObjectMapper mapper = new ObjectMapper().registerModule(new JavaTimeModule());
 
-    public UserController(UserRepository userRepo, ReviewRepository reviewRepo, TourRepository tourRepo) {
+    public UserController(AuthenticationManager authenticationManager, UserRepository userRepo, ReviewRepository reviewRepo, TourRepository tourRepo) {
+        this.authenticationManager = authenticationManager;
         this.userRepo = userRepo;
         this.reviewRepo = reviewRepo;
         this.tourRepo = tourRepo;
@@ -48,6 +48,7 @@ public class UserController {
     public ResponseObject getUser(@PathVariable("username") String username){
         ResponseObject result;
         User test = userRepo.findByUsername(username);
+
         if (test != null)
             result = new ResponseObject(true, test);
         else
@@ -56,10 +57,32 @@ public class UserController {
         return result;
     }
 
+    @GetMapping("/account")
+    public ResponseObject getAccount(){
+        WayfareUserDetails user = getCurrentUserDetails();
+        User currUser = userRepo.findByUsername(user.getUsername());
+
+        if (currUser == null)
+            return new ResponseObject(false, "No such user with username " + user.getUsername());
+
+        AccountSettingsResponse response = new AccountSettingsResponse(
+                currUser.getUsername(),
+                currUser.getFirstName(),
+                currUser.getLastName(),
+                currUser.getEmail(),
+                currUser.getPhoneNumber(),
+                currUser.getIsVerified()
+        );
+        return new ResponseObject(true, response);
+
+    }
+
     @PostMapping("/profile/edit")
     public ResponseObject editProfile(@RequestBody EditProfileRequest request){
         WayfareUserDetails user = getCurrentUserDetails();
         User toEdit = userRepo.findByUsername(user.getUsername());
+        if (toEdit == null)
+            return new ResponseObject(false, "Username not found");
 
         toEdit.setPictureUrl(request.pictureUrl());
         toEdit.setAboutMe(request.aboutMe());
@@ -75,8 +98,13 @@ public class UserController {
         User toEdit = userRepo.findByUsername(user.getUsername());
 
         request.validate();
+
+        if(userRepo.existsByUsername(request.username))
+            request.addErrors("Username already exists");
+
         if (request.hasErrors())
             return new ResponseObject(false, request.getErrors());
+
         toEdit.setFirstName(request.firstName);
         toEdit.setLastName(request.lastName);
         toEdit.setPhoneNumber(request.phoneNumber);
@@ -88,9 +116,18 @@ public class UserController {
     }
 
     @PostMapping("/account/delete")
-    public ResponseObject deleteAccount(@RequestBody EditAccountDetailsRequest request){
+    public ResponseObject deleteAccount(@RequestBody PasswordRequest request){
         WayfareUserDetails user = getCurrentUserDetails();
         User toDelete = userRepo.findByUsername(user.getUsername());
+        try {
+            authenticationManager.authenticate(
+                    new UsernamePasswordAuthenticationToken(user.getUsername(), request.password())
+            );
+        } catch (BadCredentialsException e) {
+            return new ResponseObject(false, "Invalid credentials");
+        }
+        if (toDelete == null)
+            return new ResponseObject(false, "Username not found");
 
         userRepo.delete(toDelete);
         return new ResponseObject(true, "Account deleted");
